Allow overriding StartupWMClass in .desktop file
This flag can be relevant for window grouping in app launchers (e.g.
Unity on Ubuntu). It should be set to the `WM_CLASS` attribute as
returned by the `xprop` utility.

If this is not set, it's possible that the window will not be recognized
as part of the application. If that happens, if you click on a shortcut
icon on the launcher, a second icon will pop up, instead of a dot
appearing on top of the shortcut icon.